refactor(model package): model 패키지를 domain 패키지로 이름 변경 및 generator 패키지 추가
- model패키지 하위에 있는 클래스들이 도메인의 성격을 나타내고 있어서 domain이라고 명확하게 패키지명을 변경
- generator 패키지를 추가해 generator 관련 클래스들을 해당 패키지로 이동
