+++ src/main/java/lotto/domain/lotto/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto.model.lotto;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
