+++ src/main/java/lotto/domain/lotto/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.model.lotto;
+package lotto.domain.lotto;
 
 import java.util.Collections;
 import java.util.List;
