+++ src/test/java/lotto/domain/money/MoneyTest.java
@@ -1,6 +1,5 @@
-package lotto.money;
+package lotto.domain.money;
 
-import lotto.model.money.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
